Split AppDatabase instrumentation contracts
